refactor(core): split DatabaseService into focused partial files
Improve maintainability by separating database responsibilities into dedicated partial class files while preserving the same IDatabaseService surface and runtime behavior.\n\nThe refactor organizes library cache, playback progress, playlists, and schema/migration logic into smaller units and updates project includes accordingly.
